Move DHT persistence logic into lib/ module
Increases cleanliness of index.js.
